UPSTREAM: [media] rc-main: assign driver type during allocation
The driver type can be assigned immediately when an RC device
requests to the framework to allocate the device.

This is an 'enum rc_driver_type' data type and specifies whether
the device is a raw receiver or scancode receiver. The type will
be given as parameter to the rc_allocate_device device.

Change accordingly all the drivers calling rc_allocate_device()
so that the device type is specified during the rc device
allocation. Whenever the device type is not specified, it will be
set as RC_DRIVER_SCANCODE which was the default '0' value.

Suggested-by: Sean Young <sean@mess.org>
Signed-off-by: Andi Shyti <andi.shyti@samsung.com>
Signed-off-by: Sean Young <sean@mess.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
(cherry picked from commit 0f7499fddb153a333dff3c1dc4280c178b9b5a80)
Signed-off-by: Ziyuan Xu <xzy.xu@rock-chips.com>

Note: it didn't change anything due to we didn't need the rc_dev
 Conflicts:
	drivers/media/pci/cx23885/cx23885-input.c
	drivers/media/pci/smipcie/smipcie-ir.c
	drivers/media/rc/igorplugusb.c
	drivers/media/rc/nuvoton-cir.c
	drivers/media/rc/redrat3.c