arm64: dts: rockchip: rk3528: Move usb3 otg reset to parent node
The dwc3 core doesn't reset usb controller during
system suspend if the dwc->drd_connected flag is
true which means that the controller is in device
mode or host mode. It aimed to avoid device to be
reenumerated after system resume.

However, test on RK3528 platforms, if the controller
is forced to host mode, and keep the logic power and
usb3 phy power on during suspend, after system resume,
the usb3 phy driver fail to init phy mode with the
following log:

naneng-combphy ffdc0000.phy: wait phy status ready timeout

It's because that the pipe power state isn't in the
expected P2 during initialized the usb3 phy. This patch
move the usb3 otg reset to parent node, then the dwc3
glue driver can reset the controller during system
syspend, and after system resume, it deassert the
reset and put the pipe power state in P2 by controller
hardware.

Signed-off-by: William Wu <william.wu@rock-chips.com>
Change-Id: Ieebca5c11f9ceec3ac12abecefe2db654700cbde